+++ protected/humhub/modules/post/Events.php
@@ -27,10 +27,8 @@ class Events extends \yii\base\BaseObject
 
         $integrityController->showTestHeadline("Post  Module - Posts (" . Post::find()->count() . " entries)");
         foreach (Post::find()->all() as $post) {
-            if (empty($post->content->id)) {
-                if ($integrityController->showFix("Deleting post " . $post->id . " without existing content record!")) {
-                    $post->delete();
-                }
+            if (empty($post->content->id) && $integrityController->showFix("Deleting post " . $post->id . " without existing content record!")) {
+                $post->delete();
             }
         }
     }
